+++ shoop/front/checkout/methods.py
@@ -31,10 +31,13 @@ class MethodModelChoiceField(forms.ModelChoiceField):
         super(MethodModelChoiceField, self).__init__(*args, **kwargs)
 
     def label_from_instance(self, obj):
+        """
+        :type obj: shoop.core.models.Service
+        """
         label = force_text(obj.get_effective_name(self.basket))
 
         price_info = (
-            obj.get_effective_price_info(self.basket)
+            obj.get_total_cost(self.basket)
             if self.basket and self.show_prices else None)
 
         if price_info and price_info.price:
@@ -117,7 +120,7 @@ class _MethodDependentCheckoutPhase(CheckoutPhaseViewMixin):
 
     def get_method(self):
         """
-        :rtype: shoop.core.models.Method
+        :rtype: shoop.core.models.Service
         """
         raise NotImplementedError("Not implemented")
 
@@ -129,12 +132,11 @@ class _MethodDependentCheckoutPhase(CheckoutPhaseViewMixin):
             return self._checkout_phase_object
 
         meth = self.get_method()
-        if not (meth and meth.module.checkout_phase_class):
+        if not meth:
             return None
-        phase = self.checkout_process.instantiate_phase_class(
-            meth.module.checkout_phase_class,
-            module=meth.module
-        )
+        phase = meth.get_checkout_phase(**self.checkout_process.phase_kwargs)
+        if not phase:
+            return None
         phase = self.checkout_process.add_phase_attributes(phase, self)
         self._checkout_phase_object = phase
         return phase