+++ shoop/front/signals.py
@@ -9,7 +9,6 @@ from django.core.signals import Signal
 
 # Modifying signals
 get_basket_command_handler = Signal(providing_args=["command"], use_caching=True)
-get_method_validation_errors = Signal(providing_args=["method", "source"], use_caching=True)
 
 # Completion signals
 order_complete_viewed = Signal(providing_args=["order", "request"], use_caching=True)